requirements.txt
flask
mongoengine
python-dotenv
pymongo[srv]